# Greenovation-backend-main/docs/openapiGreenovation.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Greenovation
  description: "This is the Open API documentation for the Greenovation API"
servers:
  - url: 'https://quason.tech/'
  - url: 'http://nattech.fib.upc.edu:40370/'
  - url: 'http://localhost:3000/'
paths:
...
  /events/{id}:
    get:
      tags:
        - Events
      summary: Gets the event with the indicated id
      operationId: getEventById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        404:
          description: 'No event found with this id'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /events/lletra/{lletra}:
    get:
      tags:
        - Events
      summary: Retrieves all events with the indicated energy letter
      operationId: getEventsByLetter
      parameters:
        - name: lletra
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        404:
          description: 'The letter does not exist'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /events/lletraIAnteriors/{lletra}:
    get:
      tags:
        - Events
      summary: Retrieves all events with the indicated energy letter or with a worse energy letter
      operationId: getEventsByLetterAndPrevious
      parameters:
        - name: lletra
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        404:
          description: 'The letter does not exist'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
...
  /missions/{id}:
    get:
      tags:
        - Missions
      summary: Retrieves the missions with the indicated id
      operationId: getMissionById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Missio'
        404:
          description: 'The mission does not exist'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /missions/{id}/events:
    get:
      tags:
        - Missions
      summary: Retrieves all events associated to the mission with the indicated id
      operationId: getEventsOfMission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        404:
          description: 'The mission does not exist'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /missions/lletra/{lletra}:
    get:
      tags:
        - Missions
      summary: Retrieves the missions with the indicated certificated
      operationId: getMissionsByCertificate
      parameters:
        - name: lletra
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Missio'
        404:
          description: 'The letter does not exist'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /missions/lletraIAnteriors/{lletra}:
    get:
      tags:
        - Missions
      summary: Retrieves the missions with the indicated certificated or higher
      operationId: getMissionsByCertificateAndHigher
      parameters:
        - name: lletra
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Missio'
        404:
          description: 'The letter does not exist'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: number
          example: 1
        nameCat:
          type: string
          example: "No hi ha aigua calenta!"
        nameEsp:
          type: string
          example: "¡No hay agua caliente!"
        nameEng:
          type: string
          example: "There's no hot water!"
        descriptionCat:
          type: string
          example: "Repara la caldera d'aquest veí."
        descriptionEsp:
          type: string
          example: "Repara la caldera de este vecino."
        descriptionEng:
          type: string
          example: "Repair this neighbor's boiler."
        energy:
          type: number
          example: 10
        CO2:
          type: number
          example: -5
        happiness:
          type: number
          example: 25
        money:
          type: number
          example: 50
        lletraId:
          type: number
          example: 6
        createdAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
        updatedAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
    Missio:
      type: object
      properties:
        id:
          type: number
          example: 1
        nameCat:
          type: string
          example: "Primera missió"
        nameEsp:
          type: string
          example: "Primera misión"
        nameEng:
          type: string
          example: "First mission"
        descriptionCat:
          type: string
          example: "Completa un esdeveniment qualsevol."
        descriptionEsp:
          type: string
          example: "Completa un evento cualquiera."
        descriptionEng:
          type: string
          example: "Complete any event."
        eventsForCompletion:
          type: number
          example: 1
        lletraId:
          type: number
          example: 7
        energy:
          type: number
          example: 20
        CO2:
          type: number
          example: 10
        happiness:
          type: number
          example: 100
        money:
          type: number
          example: 1000
        createdAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
        updatedAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
  examples:
...
    error404:
      value:
        message: 'Not found'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT